more files moved and re-exported
