+++ app/components/collection/index.ts
@@ -3,3 +3,8 @@ export * from "./view-menu";
 export * from "./external-search-dialog";
 export * from "./save-to-collection";
 export * from "./remove-from-collection";
+export * from "./game-sort-and-filter";
+export * from "./rate-game-dialog";
+export * from "./genre-filter";
+
+export * from "./hooks";
