+++ app/components/collection/rate-game-dialog.tsx
