+++ app/features/library/index.ts
@@ -1,10 +1,5 @@
-// Components
-import { GameCover } from "./components/game-cover";
-import { LibraryView } from "./components/library-view";
-import { GameSortAndFilterMenu } from "./components/game-sort-and-filter";
-
 // Hooks
 import { useSearch } from "./hooks/search";
 import { useFilter } from "./hooks/filter";
 
-export { GameCover, LibraryView, GameSortAndFilterMenu, useSearch, useFilter };
+export { useSearch, useFilter };
