+++ app/types/index.ts
@@ -1,0 +1,4 @@
+export * from "./games";
+export * from "./igdb";
+export * from "./playlists";
+export * from "./users";
